requirements.txt
Django==4.2.14
djangorestframework==3.15.1
django-filter==24.2
psycopg2-binary==2.9.9
djangorestframework-jsonapi==8.0.0
gunicorn==23.0.0
flake8==7.1.0
black==24.4.2
mypy==1.10.0
python-dotenv~=1.1.1
drf-spectacular==0.27.0
drf-spectacular-jsonapi==0.5.2